Move transactionInfo bottomsheet dialog to MainActivity
