+++ app/src/main/res/layout/fragment_transactions.xml
@@ -56,25 +56,4 @@
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
         />
 
-    <View
-        android:id="@+id/transactionsDim"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#99000000"
-        />
-
-    <androidx.core.widget.NestedScrollView
-        android:id="@+id/nestedScrollView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@drawable/bottom_sheet_background"
-        android:backgroundTint="?BottomTransactionInfoBackground"
-        android:clipToPadding="true"
-        app:behavior_hideable="true"
-        app:behavior_peekHeight="0dp"
-        app:layout_behavior="@string/bottom_sheet_behavior"
-        >
-
-        <include layout="@layout/transaction_info_bottom_sheet"/>
-    </androidx.core.widget.NestedScrollView>
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
